+++ src/kernel/drivers/ide.c
@@ -13,7 +13,7 @@ driver_desc_t _ide_driver_info() {
     ide_desc.is_device_driver = true;
     ide_desc.is_device_needed = false;
     ide_desc.is_driver_needed = false;
-    ide_desc.functions[DRIVER_BUS_CONTROLLER_FIND_DEVICE] = ide_find_devices;
+    ide_desc.functions[DRIVER_BUS_CONTROLLER_FIND_DEVICE] = (uint32_t)ide_find_devices;
     ide_desc.pci_serve_class = 0x01;
     ide_desc.pci_serve_subclass = 0x01;
     ide_desc.pci_serve_vendor_id = 0x00;
